refactor: Rename sessions.database to sessions.crud
