+++ backend/capellacollab/core/authentication/database/__init__.py
@@ -19,7 +19,7 @@ from capellacollab.projects.users.models import (
     ProjectUserPermission,
     ProjectUserRole,
 )
-from capellacollab.sessions.database import get_session_by_id
+from capellacollab.sessions.crud import get_session_by_id
 from capellacollab.settings.modelsources.git import crud
 from capellacollab.users.crud import get_user_by_name
 from capellacollab.users.models import Role
